untested pitch and refactor
