+++ uberduck_ml_dev/data/statistics.py
@@ -23,7 +23,7 @@ from wordfreq import word_frequency
 
 from ..text.util import text_to_sequence
 
-# Cell
+# NOTE (Sam): this file could be refactored so that it doesn't contain both speechmetrics and wordfreqencies - very different types of statistics.
 
 
 def word_frequencies(text: str, language: str = "en") -> List[float]:
